After a successful install we are no longer running @rake dradis:reset@, now we have a reset.bar script that will run dradis:reset but also will ensure that all dependencies are installed using Bundler.
However in order for this script to run, it needs to be able to find Rake and 
Bundler. If we just installed them, the PATH variable of the installer process 
would not contain the Ruby folder so we need to wrap this up in a function that
sets the appropriate PATH and then calls our reset.bat script.